Moved the contents of the Action Tab into a separate UserControl so it can be worked on outside of the whole application
