+++ FantasyModuleParser/NPC/UserControls/Action/ActionOverviewUC.xaml.cs
@@ -1,0 +1,46 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows;
+using System.Windows.Controls;
+using System.Windows.Data;
+using System.Windows.Documents;
+using System.Windows.Input;
+using System.Windows.Media;
+using System.Windows.Media.Imaging;
+using System.Windows.Navigation;
+using System.Windows.Shapes;
+
+namespace FantasyModuleParser.NPC.UserControls.Action
+{
+    /// <summary>
+    /// Interaction logic for ActionOverviewUC.xaml
+    /// </summary>
+    public partial class ActionOverviewUC : UserControl
+    {
+        public NPCModel npcModel { get; set; }
+        public ActionOverviewUC()
+        {
+            InitializeComponent();
+        }
+
+		private void LairActions_Click(object sender, RoutedEventArgs e)
+		{
+			new ViewLairActions().Show();
+		}
+		private void Actions_Click(object sender, RoutedEventArgs e)
+		{
+			new Actions().Show();
+		}
+		private void Reactions_Click(object sender, RoutedEventArgs e)
+		{
+			new Reactions().Show();
+		}
+		private void LegAction_Click(object sender, RoutedEventArgs e)
+		{
+			new LegendaryActions().Show();
+		}
+	}
+}
